requirements.txt
torch
transformers
fastapi
uvicorn
Pillow